Create project0.s
Loaded my student id to the variable id in the data section. Loaded the memory address for the id to the register $t0. Replaced the element in 0th index in the id with the element in the 1st index, while preserving the value of 0th index in the register $t1