Update project0.s
Replaced the element in 8th index with the 0th indexed element, while preserving the 8th indexed element in the register $t2 which was later used to replace the 7th indexed element while preserving its initial value in the register $t1. In the fifth line, we load the element in 6th index to the register $t2.