Update project0.s
Replaced the element in 6th index with the 7th indexed element we preserved in $t1 above, but, at the same time, we also have preserved the initial 6th indexed element in the register $t2 in the last line of the previous commit. we load the 5th index element to the register $t1 to preserve the value, while we replace the element with the element in the 6th index. Then we loaded the element in 4th index to the register $t2 and replaced the element by the 5th index element we stored in $t1 earlier.