Update project0.s
Replaced the element in 3rd index with the 4th indexed element, while preserving the 3rd indexed element in the register $t1 which was later used to replace the 2nd indexed element while preserving its initial value in the register $t2. In the fifth line, we load the element in 1st index to the register $t2 to preserve its value.