Update project0.s
Correction to the last commit: We don't need to preserve the value of 1st index since we already did that in the first line of the code. We replaced the element at 1st index with the one at the 2nd index that we preserved in register $t2. We print the result. We will see that the new id is "029689462@"