Update project0.s
After printing the first result. We again started with storing the 0th index and 1st index element (0 and 2) in the registers $t1 and $t2. We replaced the 1st index element with the 2nd index element($t2) while preserving its initial value of 0 in the $t1 register. We used the $t2 register then to load the value of the element in the 8th index to preserve its value, so we could replace it with the 1st index element that was stored in the register $t1