Update project0.s
we loaded the 7th index element (6) to register $t1 and replaced the element with the previously stored element from the $t2 (8th index element) i.e 0. we loaded the 6th index element (4) in the $t2 register and replaced 4 with 6 from the previously stored value in $t1. In the last line we load the 5th index element to the register $t1.