Update project0.s
we replaced the 9 (index 5) with the previously stored value of 6 (index 6) that was stored in the $t2 register. Then we loaded 8 (4th index) element to the register $t2. And we replaced the 8 in the index 4 with 9 that was stored in the $t1 before. Then we loaded the 3rd index element (6) to the register $t1, and replace it with the 8 that was stored in the register $t2.