Update project0.s
we loaded the elements in the 1 and 2 indices (2 and 9) to the registers $t1 and $t2. we replaced the 2 with 9. And the value 2 is preserved in the register $t1. We loaded the $t2 register then with the last element i.e. 0 and we replace it with the element 2 that was stored in $t1 earlier. Also added \n at the end of the id as it was printing in the same line