Update project0.s
Loaded the elements in the 0 and 1 indices (9 and 6) to the registers $t1 and $t2. we replaced the 9 with the 6. And the value 9 is preserved in the register $t1. We loaded the $t2 register then with the last element i.e. 2 and we replace it with the element 9 that was stored in $t1 earlier.