Update project0.s
Loaded the bytes in the 7th index (0) to the register $t1. We then replace the elements in the 7th index with previously stored element from the register $t2 (9). we then load the element @ (index 6) to the register $t2, and replace the element with the previously stored value of @ in $t1. In the last line, we load the element at index 5 i.e. 6 to the register $t1.