Update project0.s
Replaced the 5th index element i.e 6 with the value that was stored in the register $t2. We then load the element 4 (index 4) to the register $t2 and replace the element with the value 6 that was previously stored in register $t1. We then use the register $t1 to load the value at 3rd index i.e. 9 and replace its initial element with 4 (the value stored in $t2 earlier)