Update project0.s
Loaded the elements in the 0 and 1 indices (6 and 8) to the registers $t1 and $t2. we replaced the 6 with the 8. And the value 6 is preserved in the register $t1. We loaded the $t2 register then with the last element i.e. 9 and we replace it with the element 6 that was stored in $t1 earlier.