Update project0.s
Loaded the 7th index element i.e. 2 to the register $t1. We then replace the elements in the 7th index with previously stored element from the register $t2 (9). we then load the element 0 (index 6) to the register $t2, and replace the element with the previously stored value of 2 in $t1. In the last line, we load the element at index 5 i.e. @ to the register $t1.