Update project0.s
Replaced the 5th index element i.e @ with the value that was stored in the register $t2. We then load the element 6 (index 4) to the register $t2 and replace the element with the value @ that was previously stored in register $t1. We then use the register $t1 to load the value at 3rd index i.e. 4 and replace its initial element with 6 (the value stored in $t2 earlier)