Update project0.s
We loaded the elements in the 0 and 1 indices (8 and 9) to the registers $t1 and $t2. we replaced the 8 with the 9. And the value 9 is preserved in the register $t1. We loaded the $t2 register then with the last element i.e. 6 and we replace it with the element 8 that was stored in $t1 earlier.