Update project0.s
We loaded the elements '9' and '4' (index 0 and 1) to the registers $t1 and $t2. we replaced the '9' with the '4'. And the value '9' is preserved in the register $t1. We loaded the $t2 register then with the last element i.e. '8' and we replace it with the element '9' that was stored in $t1 earlier.