Update project0.s
We replaced the 5th index element i.e '2' with the value that was stored in the register $t2 i.e. '9'. We then load the element '0' (index 4) to the register $t2 and replace the element with the value '2' that was previously stored in register $t1. We then use the register $t1 to load the value at 3rd index i.e. '@' and replace its initial element with '0' (the value stored in $t2 earlier)