Update project0.s
We loaded the elements '4' and '6' (index 0 and 1) to the registers $t1 and $t2. we replaced the '4' with the '6'. And the value '4' is preserved in the register $t1. We loaded the $t2 register then with the last element i.e. '9' and we replace it with the element '4' that was stored in $t1 earlier.