Update project0.s
We replaced the 5th index element i.e '9' with the value that was stored in the register $t2 i.e. '6'. We then load the element '2' (index 4) to the register $t2 and replace the element with the value '9' that was previously stored in register $t1. We then use the register $t1 to load the value at 3rd index i.e. '0' and replace its initial element with '2' (the value stored in $t2 earlier)