Update project0.s
We loaded the elements '6' and '@' (index 0 and 1) to the registers $t1 and $t2. we replaced the '6' with the '@'. And the value '6' is preserved in the register $t1. We loaded the $t2 register then with the last element i.e. '4' and we replace it with the element '6' that was stored in $t1 earlier.